fixed customer update updated file structure
