+++ includes/Admin/CustomersPage.php
@@ -1,6 +1,8 @@
 <?php
 namespace HomerunnerBilling\Admin;
 
+use Stripe\StripeClient;
+
 if (!defined('ABSPATH')) {
 	exit;
 }
@@ -32,6 +34,15 @@ final class CustomersPage
 
 		$req_action = isset($_REQUEST['action']) ? $_REQUEST['action'] : '';
 
+		if ('delete' === $req_action && !empty($_REQUEST['id'])) {
+			$customer_id = $_REQUEST['id'];
+			$stripeclient = new StripeClient(get_option('homebill_stripe_secret_key'));
+			$stripeclient->customers->delete($customer_id);
+
+			wp_redirect(admin_url('admin.php?page=homebill'));
+			exit;
+		}
+
 		if (empty($req_action)) {
 			$customers_table = new CustomerListTable();
 			$customers_table->prepare_items();
@@ -50,24 +61,22 @@ final class CustomersPage
 				?>
 				<h1>Create Customer</h1>
 				<?php
-				require_once 'customer-form.php';
+				require_once 'views/create-customer-form.php';
 
 			} elseif ($req_action == 'edit') {
 				?>
 				<h1>Edit Customer</h1>
 				<?php
-				require_once 'edit_customer.php';
-			} else {
+				require_once 'views/edit-customer-form.php';
+			} else if (empty($req_action)) {
 				?>
-				<h1 class="wp-heading-inline">Customers</h1>
-				<a href="<?php echo add_query_arg('action', 'add'); ?>" class="page-title-action">Add New</a>
-				<hr class="wp-header-end">
-				<?php
-				global $customers_table;
-				$customers_table->display();
-			    }
-			    
-			    
+					<h1 class="wp-heading-inline">Customers</h1>
+					<a href="<?php echo add_query_arg('action', 'add'); ?>" class="page-title-action">Add New</a>
+					<hr class="wp-header-end">
+					<?php
+					global $customers_table;
+					$customers_table->display();
+			}
 			?>
 		</div>
 		<?php
