--- includes/Admin/del_csutomer.php
@@ -1,12 +1,0 @@
-<?php 
-namespace HomerunnerBilling\Admin;
-
-use Stripe\StripeClient;
-
-if (!defined('ABSPATH')) {
-    exit;
-}
-
-
-
-?>
